+++ .eslintrc.js
@@ -164,7 +164,6 @@ module.exports = {
 							'uniqWith',
 							'upperFirst',
 							'values',
-							'words',
 							'xor',
 							'zip',
 						],
@@ -357,7 +356,7 @@ module.exports = {
 			},
 		},
 		{
-			files: [ 'bin/**/*.js', 'packages/env/**' ],
+			files: [ 'bin/**/*.js', 'bin/**/*.mjs', 'packages/env/**' ],
 			rules: {
 				'no-console': 'off',
 			},