+++ lib/compat/wordpress-6.1/persisted-preferences.php
@@ -72,7 +72,6 @@ function gutenberg_configure_persisted_preferences() {
 		),
 		'after'
 	);
-
 }
 
 add_action( 'admin_init', 'gutenberg_configure_persisted_preferences' );